reorganized file structure and updated package.json
