+++ package.json
@@ -3,7 +3,11 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www"
+    "start": "nodemon ./bin/www",
+    "test": "clear; DB=test mocha --recursive",
+    "db:schema": "node ./utils/schema",
+    "db:seed": "node ./utils/seed",
+    "db:setup": "npm run db:schema; npm run db:seed"
   },
   "dependencies": {
     "body-parser": "~1.13.2",
